+++ lib/models/api/v2/entities/cashtag_entity_v2.g.dart
@@ -1,0 +1,14 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+part of 'cashtag_entity_v2.dart';
+
+// **************************************************************************
+// JsonSerializableGenerator
+// **************************************************************************
+
+CashtagEntityV2 _$CashtagEntityV2FromJson(Map<String, dynamic> json) =>
+    CashtagEntityV2(
+      tag: json['tag'] as String,
+      start: json['start'] as int,
+      end: json['end'] as int,
+    );