+++ lib/models/api/v2/entities/url_entity_v2.g.dart
@@ -1,0 +1,15 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+part of 'url_entity_v2.dart';
+
+// **************************************************************************
+// JsonSerializableGenerator
+// **************************************************************************
+
+UrlEntityV2 _$UrlEntityV2FromJson(Map<String, dynamic> json) => UrlEntityV2(
+      url: json['url'] as String,
+      expandedUrl: json['expanded_url'] as String,
+      displayUrl: json['display_url'] as String,
+      start: json['start'] as int,
+      end: json['end'] as int,
+    );