+++ lib/tweet_ui.dart
@@ -6,6 +6,8 @@ library tweet_ui;
 
 export 'compact_tweet_view.dart';
 export 'default_text_styles.dart';
-export 'tweet_view.dart';
 export 'embedded_tweet_view.dart';
+export 'models/api/v1/tweet.dart';
+export 'models/api/v2/tweet_v2.dart';
 export 'on_tap_image.dart';
+export 'tweet_view.dart';